+++ src/info.py
@@ -1,0 +1,14 @@
+import random
+import math
+
+#### All informations about the Python Module
+
+def info():
+    print("Rebirth | Python Module")
+    print("- A simple python module using simple functions")
+    print("- Useful for some things like vector calculations")
+    print("- Less useful but still good like pi_gen")
+    print("- Made for fun.")
+    return info
+
+# rebirth.info()
