+++ src/ping.py
@@ -1,0 +1,7 @@
+import random
+import math
+
+#### Ping
+def ping():
+    print("Pong(1)")
+    return ping
