+++ src/rand.py
@@ -1,0 +1,11 @@
+import random
+import math
+
+#### Generate a random number
+
+def rand(min1, max1):
+    rand = random.randint(min1, max1)
+    return rand
+
+# d = rebirth.rand(1, 10)
+# print(d) random number.
